Extracted a class from the handler. decoupling
